+++ src/api/process.rs
@@ -1,8 +1,8 @@
 use crate::api::syscall;
 
 pub fn spawn(path: &str) -> Result<(), ()> {
-    if syscall::info(&path).is_some() {
-        syscall::spawn(&path);
+    if syscall::info(path).is_some() {
+        syscall::spawn(path);
         return Ok(());
     }
     Err(())