+++ src/usr/host.rs
@@ -124,7 +124,7 @@ impl Message {
 
 fn dns_address() -> Option<IpAddress> {
     if let Some(servers) = usr::net::get_config("dns") {
-        if let Some((server, _)) = servers.split_once(",") {
+        if let Some((server, _)) = servers.split_once(',') {
             if let Ok(addr) = IpAddress::from_str(server) {
                 return Some(addr);
             }