+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+- Improve dhcp command (#335)
+- Rewrite network interface (#334)
+- Add CSI sequence for enabling or disabling echo to userspace (#333)
+- Refactor code (#330)
+- Improve FUSE driver with write and delete (#292)
+- Fix device reading (#329)
+- Add a reboot command (#328)
+- Fix issues with process alloc (#327)
+
 ## 0.7.1 (2022-04-10)
 - Add 2048 game ([#295](https://github.com/vinc/moros/pull/295))
 - Add Box to process data ([#306](https://github.com/vinc/moros/pull/306))